SurfaceDescriptionFunction starts to work.
