+++ com.unity.shadergraph/Editor/Data/Util/GraphUtilForVFX.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 130a64d32c054c5fb51b1e5e13f93fc2
+guid: 922cb4ade3e5dfa4f932006e1caf795a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
